Refs #10392 Sort runs list in ReflSearchModel
This commit fundamentally changes the behaviour of ReflSearchModel.
Instead of merely wrapping a ITableWorkspace, it copies the data it
cares about out of the workspace and then forgets about it. The main
reason for doing this is the lack of sorting functionality in the
TableWorkspace class. Weighing things up, the simplest way forward is
simply to look after the data we care about ourselves and manage it
internally.