Refs #10392 Fix substr exception on RHEL6
What was the cause of this mysterious bug? An older version of
boost::regex. It was matching successfully but not bothering to fill out
the match object, so the position parameter being passed to
string::substr was just uninitialised stack. Telling boost to reset
branching between each pattern solves this problem.