Refs #10263. Adding new algorithm PoldiPeakSummary
PoldiFitPeaks1D produces a summary table with peak information. Since this will re-occur in the 2D-fit and also in version 2 of the algorithm, this is put into a separate algorithm. It's very small and it will be much easier to deal with possible changes to the output format of poldi peaks.