Refs #9952 Rework spectrogram source saving
Since v3.2 and prior crash when dealing with spectrograms, we don't have
to worry about compatibility, so we're allowed to deal with this a
little more simply.

If the spectrogram's based off a matrix, we specify the matrix by its
name. If the spectrogram's based off a workspace, we specify the
workspace by its name.

Load and save accordingly.