refs #10820. Use original dim name
The problem at the moment is that if you execute plotMD, and what should
be the equivalent (ConvertToMatrixWorkspace followed by plotSpectrum), the
x-axis labels are different. plotMD takes the input from the original
workspace, which is better, where as the latter approach is providing
arbitrary labels. If the transform to the original is possible, and the
original workspace has been provided, then that is where the dimension
information is now taken from. Unit tests adapted to show this.