refs #10846. Removal listening
Added the functionality to automatically stop managing peaks workspaces that are removed in the ADS. This is done slightly differently from the update functionality because we already have the gui commands for removing the peaks workspaces via buttons in the PeaksViewer. Therefore the notifications go down through the PeaksViewer rather than up through the CompositePeaksPresenter, as in the case for the updates.