Refs #10263. Removing "recursion borders" from PoldiPeakSearch
After careful consideration and inspection of the code and what it does and does not find, I came to the conclusion that this is not needed at all. It caused some strange problems in the past with debug-builds and so I decided it has to go. In fact, if anything, it was cutting down the searched parts of the vector.