refs #10952. Observe rename events.
We need to observe and act on the special case of a rename event. The rename algorithm execution not only results in the key change in the ADS dictionary, but also the value change (workspace deep-copy). We therefore only have the old-name as a constant reference to use for find-removal. Only the PeaksWidget has this information so we go top-down to purge.