refs #10904. IntegratePeaksMD algorithms done.
IntegratePeaksMD v1 and v2 write out integrated peak info in the form of PeakShape

PeakShapeSpherical adapted to write out background outer radius and background inner radius as well as the peak radius. Factories also updated for this, unit tests added to and updated for this.

I got a little worried that someone may change the integer values agaist the enum entries of SpecialCoordinateSystem. So I put in place a unit test to sanity check those.