Refs #10555. Fixed a critical error on run and detid.
A critical error is fixed on the run ID and detector ID in the
MDEventWrokspaces.

More unit tests are added to ensure that the rotation of the detectors
are reflected in the MDEventWorkspaces.

Property InitRunNumber is removed and replaced by using Pt. value in the
SPICE file.

Documentation is modified for better understanding.