refs #11056. CreateSampleWorkspace fixed.
CreateSampleWorkspace does not set up a ReferenceFrame which is consistent with the actual layout of the instrument.
I don't know how this probelem got in there in the first-place, or why it hadn't cause problem up to now! Peak calculations require that the reference frame is known so that we can dot product with the beam direction in order to determine dQ.

Fixing this should now make the doc-tests pass.