refs #9052. Upgrade CreateSampleWorkspace
Add options to move the banks and set the pixel spacing for the fake instrument.

This is an important feature of working with SCD data, where we want to move the bank in such a way that we have coverage for specific reflections. I've unit tested the two new properties I've added as part of this. The default behaviour will be exactly the same as before.