Refs #11360 - Coerce Sophos into playing nice.
Sophos is complaining about unistall.exe again for some reason.  It's
possible that it's quite flaky about changes to XML files, and the
only XML file that we changed in this branch was the LARMOR param file.

Making a trivial change to a comment to try and motivate a different
result.