Re #11508 Addressing the reviewer comments
1) Tooltip name changed to short description

2) Why simply not use m_description = source?
because for strings m_description = source compiles into something like
if !(&rhs == this)
   lhs.assign(rhs)

3) The names of these methods don't properly reflect what they do. For example, Component::getDescription would imply that the description is about the component, not the parameter...

The name means exactly this. Component::getDescription provides the description for a parametrized component. By current agreement (at least enforced by unit tests), a parametrized component would have a parameter with the same name. If it does not, empty description will be returned. If one sets description for a component, a parameter with component name will be created if not already present. I do not need this feature but decided that it would be nice to have as implementation is really trivial, behaves similarly to all other stuff in components and may be useful to somebody.

4) From my perspective, would these not be a better set of methods:  getParameterDescription, getShortParameterDescription,setParameterDescription...

Done. Except I've used getParamDescription etc instead of getParameterDescription -- it is as descriptive as suggested but saves a bit of typing.
setParamDescription is not implemented. It is not entirely trivial and I do not need it so let's leave it to somebody who(if) actually needs it.

5) I think the description would be more sensible as simply the value of the description element rather than an attribute, i.e....

I make it the same as 12 other (parameters? attributes? components?) parts of the Instrument tree are. Though doing it the suggested way would be better, all other parts are implemented this way. Introducing new entities would be a bit too much for such a small change to the instrument tree. But it its really desired we may create a separate ticket to do such change with this and number of other instrument components. I only suspect that the gain is not worth the efforts, though would be nice.

I think the changes made to the ticket address all remarks.