Minor changes to how data is parsed
Change how the TOPAZ listener calculated the indexes in the event list
to work around a problem with event_catcher sending out bad data.

Add a log message for cases where we have to discard the events in the
packet.

Also, we only need to add to the PROTON_CHARGE time series once per
pulse, not once per event.

Refs #11315