Update NormaliseByCurrent-v1.rst
The Windows Doc test has been failing repeatedly on this doc-test. I've run this python section manually in mantidplot and it seems ok, and there's nothing in the code that reveals otherwise. As this has been causing problems for ALL PRs, I'm temporarily disabling this section of the doc test. If I can establish what the real problem is with this, I'll revert this commit.