refs #11872. New form of delete peaks.
New form of deletion of peaks based off a cursor region. Uses PeaksOnSurface.
Determination of Cursor position in Natural coordinates
Plot current axis needed to be passed through.
Unit tests updated and modified. All passing.

There are still some issues with the cursor and the deletion modes that will need to be fixed.
* Cursor doesn't seem to properly go away.
* After delete update on the PeaksOverlayMultiSphere causes a crash when driven from the IntputController
* Need to persist the slicing tool because the PeaksOverlayMultiSphere is thrown away and created afresh after the PeaksWorkspace has been updated.