refs #11872. Fix crash.
The crash came about because the occupancy settings defer to the first PhysicalPeak. If we have deleted all peaks, then we have an invalid read. To fix this, I have cached the occupancy settings on the widget, that we we correctly keep any set occupancy values even if there are no peaks present.