Fix extraction of C++ Workspace types from Python
By default boost will try to construct a new pointer type when extracting
a C++ type from a Python object. We need to ensure we get back the same
pointer object so we use a reference type in ExtractWorkspace. This then
requires that all return types that convert to a Python object return
a common Workspace_sptr or the reference extraction fails.
Refs #10649