Re #12897 Test for size-1 input to convertToBinBoundary().
As fas as I understand there is no definite way to deal with a bin
center vectro of size 1 --- we cannot compute a bin width. However, the
current implementation is broken: it depends on unitialized values and
yields things like an upper bin edge that is smaller than the lower
edge.
This test test (somewhat) arbitrarily for a bin width of 1.