Re #12917 Protected copy const. / assigment op. in Workspace childs.
Workspace is polymorphic, so in general we do not want any public copy
or assignment. In some cases this required moving the user defined copy
constructor out of the public area, in other cases it meant adding it in
the protected section, to avoid auto-generation by the compiler.

Note that we make the copy constructor protected and not private. This
is in view of the upcoming introduction of a clone() method, which will
need to call the copy constructor of the base class.