Re #12687 Cleanup/unification in CloneWorkspace algorithm.
Using the polymorpic clone, looking at the inheritance structure, code
in CloneWorkspace could be unified. PeaksWorkspace is a child of
ITableWorkspace, so Peaks* and Table* cast could be combined.
EventWorkspace is a child of MAtrixWorkpace, so they could also be
combined.
The remaining special case is due to MDEventWorkspace, which still has
some special code on CloneMDWorkspace.