Re #12687 ISplittersWorkspace is not a Workspace anymore.
The reason for this change is related to the diamond inheritance that we
had previously for SplittersWorkspace. VS2012 did not manage to compile
this in combination with the virtual doClone() function with covariant
return types.
Just like IMaskWorkspace, ISplittersWorkspace is now not a Workspace
anymore, and thus the diamond is gone.