Update ADARA files with latest version
Import the latest ADARA files over from the ADARA git repo and make the
necessary mods to work with the Mantid code.  Specifically, we needed
the updated parser and new packet types.

Also fixed a pretty serious bug in the handling of the annotation
packets.

This code seems to work with old ADARA data, but I still need to test it
against an ADARA stream that includes the new packet types.

Refs #13200