Edited IDF for OFFSPEC
Offspec can now find the information associated with slit1,2,3,4.
There was an error in definition of slits in the XML which is now corrected.
As this file was given by an instrument scientist and there are existing system tests for this file there are no more appropriate tests that can be ran for this.
Refs #11806