Fix up ADARA packet unit test
The recent changes to the ADARA* files requires a few minor changes to
the unit test.  Specifically, veto() is now vetoFlags() and
bufferParse() now takes 2 parameters.

Also added a using statemnt to clear up a bunch of -Woverloaded-virtual
warnings.  (The code worked fine without it; this just makes the compile
cleaner.)

Refs #13200