Periodic Table now loads into CustomDialog
Colours and Borders of buttons are set, but still need to find out
why the horizontal expansion of the window has no limit.
All of the styles for the PushButtons are now done through
setStyleSheet in the cpp file using a QString as the style sheet.
This needs to be done as the style sheet for the button seems to override
whatever style you implement through the QPushButton member functions in the cpp file
Refs #13146