Dialog for GetNuMegMuonicXRD links up with python alg now
Elements that are select from the periodic table widget are put into
a comma separated list which then used as the elements argument for the
python algorithm.

A text box for YAxisPosition is also included in the dialog and this
value is now passed as the YAxisPosition argument for the python algorithm

Still need to work on the layout of the dialog, code formatting and cleanup
and freeing up memory when the dialog is closed.

The algorithm works and produces both a single and group workspace when
ran through the dialog.

Refs #13146