Changes to custom Dialog
The custom dialog now only enables the elements for which we have data for
as well as validating the input given by the user. If no elements are selected
or no YPosition value is provided then the algorithm will not allow the user
to press Run.

The Dialog works properly now and will produce the correct workspace(s) from the input provided by the user

Refs #13146