Changes to functionality of PeriodicTableWidget
PeriodicTableWidget now has methods for enabling buttons by a string
and also by a group. There is also a method for disabling all buttons.
Need work more on the functionality and provide a way of enabling all buttonsand also disabling a button by name to add converse functionality to PeriodicTableWidget

Refs #13146