Included an OutputWorkspace property and started unit test work
Now if GetNegMuMuonic algorithm is run from the python interpreter
an OutputWorkspace name must be provided. This output workspace will
be a group workspace containing at least one workspace with the
muonicXRD data in.

Started writing tests for the algorithm, checking if the workspaces produced
by the algorithm contain the correct data and such.

Refs #13262