Refactoring of creating table and writing row data
There are now methods that create the table for writing to and also
for writing each row of that table.

Fixed Y values being read from input workspace to now read all indices instead
of just the zeroeth entry. Needed to do this because we cannot assume that the
workspace is uniform and it might be a ragged workspace.

Refs #13410