refs #13466. Lock aspect ratios
Aspect ratios should be locked for Q dimensions. If the x and y dimensions of the plot are both Q dimensions, we lock the aspect ratios. Otherwise we don't. Need to add menus for this feature.

Had to fix a problem with the regex. boost::regex_match matches the entire string. Used boost::regex_search instead.

Introduce menus for selecting/unselecting fixed aspect ratios.