refs #13989. Model takes transfer strategy.
The catalog searcher used to pre-filter out anything but raw files. This is a problem because our filtering depends on the transfer strategy. Instead. We defer the filtering util the model is displayed, and we use the TransferStrategy to tell us which files we can use in our transfer list.