Refs #13906. Small rearrangement.
Removing the second call to Peak::findDetector() revealed what I believe was a bug in PredictPeaks due to the change in Peak that introduced the implicit call to findDetector when Peak is created.

In the first call to findDetector, the correct scattered beam direction (derived from Q) is used and potentially a detector is found. In the second call, instead of using the same beam direction, a new beam direction is calculated using Peak::detPos, which may intersect with a different detector. Depending on the beam direction, subsequent calls to findDetector will shift the peak further.

The immediate solution for PredictPeaks is to not call findDetector a second time. For a more permanent solution, a ticket addressing this will be opened.