refs #141219. Fix regression test.
When sub-ids are combined, we need to know the absolute position in the final table where we are going to add the runs. The issue before was that each set of grouped measurements set up it's own sub_id list, which was fine, but it was recording the index into that sub_id list rather than the index into the final output list of rows. That meant we where adding subids in the wrong place.