Empty values in algorithm history Re #14207
If a property is not set for an algorithm and its default value is set
to be an EMPTY_INT/DBL/LONG, do not show the internal representation of
this value to the user in the Workspace History dialog. Instead, show an
empty property.

Add a method to PropertyHistory to test if property is set to an
EMPTY_XXX default value. Use this test in AlgorithmHistoryWindow.