Re #10329 Fix Load Current for ARGUS at ISIS
ARGUS was previously treated as a special case, however found during
testing that current data would not load.
It seems that the data is now treated the same as the other instruments,
i.e. with \\\\ARGUS\\data\\autosave.run pointing to it.
Changed and tested that it does load yesterday's data.