Re #14460 Refactor plotting code in MuonAnalysis
Refactored python code to use the workspace's title directly rather
than passing the title in as a string and having to escape the single
quotes. Tested with the original WS giving error in the issue.

This will work fine unless the actual workspace name has a single
quote in it (unlikely)