Re #10329 Use mounted SMB and fix ARGUS access
Once SMB share has been successfully mounted for non-Windows machines, it
can be used.

On all platforms, accessing ARGUS (ISIS) current data was failing. The
special case (data on \\ndw828) no longer applies and now it works in the
same way as the other instruments, so removed the special case.