Disable static linking to libstdc++ on RHEL6
MantidPlot no longer starts when this is combined with more recent changes to the code. It also turns out that the gcc provided with devtoolset includes modifications to statically link newer libstdc++ symbols that are not present in the  system version so that the binaries will run on legacy systems without the devtoolset.