Re #14390 LoadingPresenters take ownership of view - change signature
The xxxLoadingPresenter objects take a raw pointer to a view in their
constructor, and then delete it in their destructor. This is confusing
to the caller. Replace the raw pointer with a unique_ptr - the presenters
now take ownership of a unique_ptr and this is much clearer. Updated all
calls in VatesAPI and tests.