Re #10057 Get most recently modified nexus file in directory
If "auto" is selected, set last run to be the most recently modified .nxs
file in the directory

At the moment, doesn't check the prefix, but it does check the extension
is .nxs (case-sensitive).

Make sure that both paths are sent to the algorithm in the native format.
(On Windows, m_view->firstRun() returns a path with forward slashes, and
PlotAsymmetryByLogValue then thinks the files are not in the same directory
even when they are).