Re #10057 Fix re-entrancy problem adding/deleting multiple files
To replicate bug: Use auto mode to load whole directory. Delete 2 files
from middle of list. Error "invalid string position" is thrown from the
PABLV algorithm because updateFilesFromDirectory is called twice in quick
succession and the algorithm is re-entrant.

Fixed by a 'changed' flag. When change detected, set the flag. Every
second, flag is polled and, if set, data is reloaded and flag is reset.