Re #13916 Change MuonLoad's input properties
MuonLoad is going to take over the functionality of MuonCalculateAsymmetry.
Looking at where it is called in the MuonAnalysis interface, there are
three use cases (after LoadMuonNexus has been called):
- correct and group
- change bin offset, crop, rebin, calculate asymmetry
- both of the above
and these are the three choices for the "mode" input parameter.

MuonLoad now takes a workspace so that the work done in MuonAnalysis can
be done here instead. This will require calling LoadMuonNexus beforehand
both in MuonAnalysis and in MuonSequentialFitDialog. (And MuonLoad should
probably be renamed! MuonProcess?)