Re #13916 Rename MuonLoad to MuonProcess as it doesn't load any more
The name of "MuonLoad" is now very confusing as it no longer loads data
from a NeXus file. Change it to "MuonProcess" as this better reflects what
it does.

This is also a good thing because the algorithm has changed quite a bit,
and so changing its name will avoid confusion with the old one.