Re #14048 Enable autoscaling in ALC (and other) data plots
Set autoscale to on in setDataCurve, and override boundingRect() in the
ErrorCurve so that a new rect can be returned, taking account of the error
bars.

Qwt's autoscaler uses boundingRect(), which by default returns an invalid
rectangle (no autoscaling) but can be overridden to provide the bounding
rectangle of all points and errors.