Re #14739 User selects which spectrum to plot from each workspace
In the case of multi-spectral workspaces in the workspace group, user is
offered the choice of which spectrum index to use to generate the plot.

The waterfall and "Plot All" options are disabled as they make no sense
here. Despite this the user could still type a range of numbers; in this
case we just use the first (the index used is displayed in the title of
the graph so it is clear what has happened).