Re #14739 Refactor code in MantidWSIndexDialog into a reusable widget
Take the code out into a widget that can be reused in other dialogs, and
leave the dialog itself as just a container for this widget.

The point of this is so that the new dialog for surface plots can use
this widget too, but also have other things in the dialog.