Re #14739 Cache value to speed up plotting
Profiling reveals plotGraph3D is the bottleneck. Most time is spent in
Graph3D::addFunction and MantidMatrix::cell. cell(i,j) is called four times
with the same values each time through a nested loop, so changed this to
calculate value once each time and cache it.

(Most of the code changes in this commit were forced by clang-format
rather than actual changes to the code).

Now the main bottleneck is Graph3D::addFunction, in which Axis::length is
the culprit.