Re #14739 Contour plot: keep plot when workspace deleted
Previously, the workspace used to generate the contour plot had to stay in
the ADS, and the plot disappeared when it was deleted. This also meant
that only one contour plot could be generated per Mantid session! Clearly
this was less than optimal, so now the plot is created and then copied to
avoid this problem.

Delete the existing layer before copying the new one in