Re #14818 SliceViewer: don't update colour scale for live data
If the workspace being looked at is live data, don't update the colour bar
every time it is reloaded - only do so the first time.

The SliceViewer needs to know which workspaces contain live data. It does
this by checking for all running instances of MonitorLiveData when opened,
and storing their output workspace names. This guards against the case
where more than one live data reader is active, and the slice viewer is not
using the most recent - it will still be able to tell that it's live.