Re #14840 Show fitting progress in ALC interface
The Muon ALC interface's fit did not show a progress bar, so run the fit
algorithm asynchronously so that progress is shown. The busy wait is
necessary so that the GUI doesn't freeze in the meantime - this seems to
be the standard way of doing this in the codebase at present.