Re #14818 Use global autoscale property in color fill plots
If the global "autoscale" property is set to off, don't autoscale the
color bar axis in color fill plots. If it is on, autoscale as normal.

This is the property under View > Preferences > 2D Plots > Autoscale.
Must propagate it through to the Spectrogram with the color bar axis
so that it can use this property.

Use case: color fill maps of live data with user-set color maps should not
reset the scale every time new data comes in