Re #14739 Only show logs that appear in all workspaces in group
Logs to plot against must be single-valued, numeric, and appear in all
workspaces in the group. This is done by constructing a hash table of
(single-valued, numeric) log names versus how many workspaces they appear
in. If the count is equal to the number of workspaces then this log name
is added to the combo box.