Re #15006 Guard against null pointer
When using the log scale in VSI, a crash was observed. This was found to
be caused by a null pointer to a lookup table.

In other places in the same source file, the pointer returned from
getLookupTable() is tested against nullptr before it is used. Add code
to do the same here so that the crash is prevented.