Refs #15117. Change how import is checked
Instead of not registering the algorithm at all, the import is now checked when the algorithm is executed and produces an error that guides the user to the right python package to install. The intention of this is to enable building of documentation. Otherwise users wouldn't have any way of knowing that this algorithm exists.