Re #15066 Use extracted methods to get grouping from instrument
If the grouping is not provided, read it from the IDF using the method
just extracted from MuonAnalysis.

If instrument is MUSR, this is not possible as we don't know whether the
main field was transverse or longitudinal (this information is in the
NeXus file but is not in the workspace), so for MUSR grouping must be
specified.