Minor changes to Kernel::Matrix re #13998
- Fix setMem when setting for the first time

If setMem is called with number of rows, columns the same as existing,
nothing is done as the memory is assumed to already be set. Test if the
memory is set and, if it is not, set it for the first time.

- Rename functions to better reflect what they do

The functions "fDiagonal" and "bDiagonal" were very confusing, with the
comments talking about forward and backward diagonals. What they actually
do is nothing to do with that - they form a diagonal matrix from the given
vector and pre- or post-multiply the given matrix by that.

They are not used anywhere in the codebase (besides their own tests) and so
are a candidate for deletion anyway.

- Correct comment to match what function actually does (Ssize)

- Remove unused method (Faddeev)

- Throw instead of returning error code (GaussJordan)

note this function is not used anywhere