Re #15294 Using GeometryInfo in RemoveBackground.
Note that an ASSERT for a certain exception was removed from a test:
This was not actually testing what it claimed, but a later exception
(the same one from the subsequent ASSERT). There was no easy solution
for creating that failure. In fact, it is unclear to me when the unit
pointer would ever be zero, given that it is initialized in the Axis
base class --- maybe a check for an Empty unit would be better?