Add include for definition of MantidVec
Previously was relying on a nested include for the definition of MantidVec, which is defined in cow_ptr.h.
In the meantime this has been reorganised, so this branch wouldn't build after merging into master.
Include cow_ptr.h explicitly here.

It needs to be in the header file as MantidVec is used in the declaration of the function as well as the .cpp

re #15264