Simplify setting of group/pair to plot
re #15457

The aim: to simplify where "what's being plotted" is changed

The UI (drop-down box on the Home tab) was previously changed
from many different places.

Refactor more functions into API::Grouping and API::GroupingLoader and
make a class out of IO_MuonGrouping.

Now the only place where the drop-down box is changed is in MuonAnalysis,
in the setGroupOrPairToPlot method. The UI is always up-to-date with what
is going to be plotted.