Removed outdated assert in RemoveBackgroundTest. Re #0.
This assert applied only when GeometryInfoFactory would fail to
construct when no source or sample was set. This is no longer the case
after recent changes, so this is removed from the test.
If the algorithm is *ran* with an invalid instrument it should still
throw, however this is tested by the tests of GeometryInfoFactory and
not duplicated here.