LoadMuonNexus v1: use IDF if no grouping in file
If the NeXus file has no grouping or zero grouping, load the grouping
from the IDF. If this fails too, use a dummy grouping. In either case,
log a warning message.

This enables the use of such files in the ALC interface, and means the
behaviour is the same as in MuonAnalysis.

re #14085