Correct X and Y axis labels on plotting from table
Take the X and Y column names from the DataCurve.
They are in the form "Table-1_axisTitle" so split on '_' to get the titles.
Only split the axis labels if they are of the right form.

Note that the legend string is left as "Table-1_f0.A0" or similar - this
is because you might plot data from two tables on the same graph, and
then the legend would rightly say "Table-1_f0.A0" and "Table-2_f0.A0". If
we removed the Table part it would be confusing.

re #15226