Added Scale and Rebin to RRO
Two post-processing steps, Scale and Rebin, can now be applied to the workspace
when the correct parameters are supplied by the user. There has been unit tests
added to reflect these changes. Documentation will need to be updated to explain
the new properties and how they can be used by the user.

Refs #15865