Don't initialize PeakPickerTool from a fit curve
This caused the bug in the issue when a graph had:
- curve 0: fit to dataset A (old)
- curve 1: data from dataset B (new)
and the PP tool initialised from dataset A, meaning that the workspace
A was the one shown in the "Data" tab.

(Rationale:
It doesn't make sense to initialise the PP tool from a fit.
It's a tool for fitting data, and why would you fit a fit curve?)

re #16017