Replace busy wait with signal and slot
The sequential fit dialog runs the file search in case the user
has edited the run input box. We need to wait for the search to
complete before getting the filenames.

Previously this was done in a busy while loop using
QApplication::processEvents(). This has been changed so that the
fit is only started once the search reports that it is done.

(We use a new signal fileInspectionFinished() because the existing
fileFindingFinished() is emitted "too soon", i.e. when the thread finishes
but while its member variables are still being accessed).

re #9963