Disable choice of spectrum number if data is single-spectrum
When plotting a surface or contour plot from a WorkspaceGroup on
right-click, the dialog offers a choice of spectrum number (or workspace
ID) along with log values. Disable this choice for single-spectrum data.

(Note this doesn't affect the regular "plot spectrum" option - there
the dialog isn't shown at all for single spectrum data. Here we still
need to show it to offer the log value/axis name choice).

re #14964