When new workspace loaded, set group/period boxes
When a new workspace is loaded in the Home tab and the user switches
to Data Analysis, set the appropriate group/period boxes checked in the
data selector.

We can work out which boxes to tick by parsing the workspace name.

Add test to make sure this happens. It should only occur when a *new*
workspace is loaded, not if the user switches back and forth between
tabs without changing anything.

re #15518