Re #16282. Added a simple Workspace factory function.
By using this we can avoid unnecessary casting when we need, e.g., a
Workspace2D. The reason why this is relevant now is that the setters for
Histogram are not forwarded to the MatrixWorkspace interface, since they
would not make sense for all child classes (in particular
EventWorkspace). That is, when we want to set values for Histgram Y and
E, we need to work with Workspace2D instead of MatrixWorkspace.