Assume files go in run number order
When working on a remote directory with lots of files, ALCLatestFileFinder
is unacceptably slow because it queries every file for its last modification
time. These times can be inaccurate anyway if the archiver has misbehaved
or if files have been copied in a batch.

Change the way this works to use run numbers. It's ok to assume the files
were created in run number order (confirmed with scientists).

re #16373