Fix grouping error when loading old muon files
Error thrown for multi-period files that have number of grouping entries
= number of spectra. Deal with cases like this.

In the case of the test file, this also fails and there is no IDF for DEVA,
so it falls through to getting the dummy grouping. Add a fix because
the grouping loader doesn't know the correct number of detectors.

re #16396