Hack to allow MuonAnalysis to load DEVA data
DEVA was an old muon instrument at ISIS. There is currently no IDF for
this instrument and so it isn't supported by MuonAnalysis.

Add a special case to ignore the warning, and deal with zeroPadding:
now old DEVA data can be loaded and fitted.

In the long run it would be better to add this instrument to Mantid.

re #16396