Fix bug where GUI was updated when it shouldn't be
The data selector GUI was being updated whenever a new workspace was
loaded. It should only be updated when a new workspace is loaded **in the
Home tab**. If we have created some analysis workspaces to add to the list
of workspaces to fit, don't update the GUI.

NB This assumes that the workspace loaded first on the Home tab will
always be fitted - this is the intended use case from scientists. But as it
stands, you can de-select the options on the GUI and find it still gets
fitted, which is confusing.

re #15518