Re #16734 Make the generic presenter push the new actions
Also create slots for when the actions are triggered. In addition, I removed the 'Prepend Row' action. The reason is that there's already an action to append rows, and if I kept 'prepend rows' I would need the equivalent for 'prepend group', which would make three new actions, which is prehaps too much