Enable drag to resize function browser, data selector
Put the "additional widgets" in a QSplitter so their relative sizes can be
adjusted. This is useful when a complicated composite function with many
parameters is used, so that more can be seen without needing to scroll the
function browser.

Use stretch factor to ensure that the additional widgets get any extra space
when the window is resized.

re #15518