Disable "Browse" button for runs on data selector
Use case is that first run will always be the run selected on front tab, so no
need for this button as user will always type runs in.

Allowing the button creates problems because then getRuns() will return
a comma-separated list of filenames which will need parsing - disabling
the button sidesteps this problem!

re #15518