Re #16902. Removed optimization that may actually make things slower.
The optimization that was removed avoided a sqrt computation. Compared
to the rebin cost this is probably a small cost. Since Y and E are
typically accessed both, the effect of this optimization is that the
rebin has to run twice, once to get Y, and once to get Y as an
intermediate result for E.