Fix bug creating MultiDomainFunction from string
The createComposite function is recursive. If a child function has a
"parent attribute" (one that starts with $, like $domains) then it is
passed up to its parent and applied there.

In this case we had:
"composite=MultiDomainFunction,NumDeriv=true;(name=FunctionFactoryTest_"
        "FunctA,a0=0,a1=0.5;name=FunctionFactoryTest_FunctB,b0=0.1,b1=0.2,ties="
        "(b1=0.2),$domains=i);(name=FunctionFactoryTest_FunctA,a0=0,a1=0.5;"
        "name=FunctionFactoryTest_FunctB,b0=0.1,b1=0.2,$domains=i);ties=(f1.f1."
        "b1=f0.f1.b1)"
so the domains attribute was passed up to the parent CompositeFunction, but that
function doesn't have that attribute - it needs to be passed up again to the
grandparent MultiDomainFunction!

Now the new test passes.

re #16942