Fix crash loading old MUT data into MuonAnalysis
LoadMuonNexus1 looks for number of good frames (goodfrm) in
run/instrument/beam/frames_good for single-period data. This particular
old MUT file didn't have such an entry, and the absence of goodfrm in the
workspace's run object caused a crash later on.

If frames_good isn't there but frames is, use frames and assume they are good.
Log a warning so that the user is kept informed.

re #16723