Fix use of EvaluateFunction from GUI dialog
The dialog called addWorkspace again before executing the algorithm. Because
m_domainCreator was not null on the second call, its type was checked, but it is
not MultiDomainCreator, it is FitMW because the function has only one domain.

Check number of domains in the function before checking whether the creator is
a MultiDomainCreator.

re #17092