Re #17061. Removed bad cast from int64 to int32, fixing OSX issues?
Previously the index type was size_t, i.e., also 64 bit, but actually
contained the spectrum number, which was 32 bit, so casting to an int
did not cause trouble. Now we are using the this pointer as index, and
the values exceed 32 bit. This might explain the failing unit tests on
OSX (clang).