Use point data in FFT, output point data workspace
Whether the input is histogram or points, we use points internally and
output the FFT as points - this is the only way that makes sense.

Using histograms can give errors, for example a symmetrical function does
*not* give an entirely real transform with histogram data, but it should!

The user can manually convert the output of FFT to histogram using
ConvertToHistogram, if that is what they want.

re #16405