Detector ID for muon Nexus v1 files
For muon Nexus v1 files, there is always a one-to-one mapping between
spectrum number and detector ID.

Set detector IDs in workspace so that they are correct in the case that
not all spectra are loaded, e.g. load spectra 1-5, 8-10 then detector IDs
shouldn't be 1-8 but 1-5, 8-10 also.

Add a note to documentation.

re #16628