Limit scope of changes to global config option
For muon fits, scientists want the option "curvefitting.peakRadius" set to 99.
The scope of this change should be limited to the times that muon fits take place,
i.e. while the data fitting tab is open.
This avoids messing with fitting in other areas of Mantid.

(During this time, the muon interface's FitPropertyBrowser takes over as Mantid's default too).

re #0