Fix invisible graphs when more than 6 subplots
There was a problem displaying the MultiLayer when it had more than
6 subplots. The problem was that the width took a negative value.

This commit is a hack to just take the sign of the calculated width.
This seems to do the right thing and the plotted result looks good.
TODO: investigate why this works, and why w took a negative
 value in the first place!

re #9239